News explorer hub/requirements.txt
streamlit==1.38.0
requests==2.32.3
google-generativeai==0.8.2
python-dotenv==1.0.1